feature/REG-37
- change types in database and casts of models for coins volatile data from decimal to float for easier calculations and number representation in JSON and Vue.

+++ database/migrations/2025_01_25_203805_create_portfolio_coins_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('portfolio_id')->constrained()->cascadeOnDelete();
             $table->foreignId('coin_id')->constrained()->cascadeOnDelete();
-            $table->decimal('quantity', 18, 8)->default(0);
+            $table->float('quantity', 18)->default(0);
             $table->timestamps();
         });
     }
